+++ CHANGELOG.rst
@@ -1,10 +1,33 @@
 Changelog
 ---------
 
-20.0 - `master`_
-~~~~~~~~~~~~~~~~
+20.0 - master_
+~~~~~~~~~~~~~~~~~
 
 .. note:: This version is not yet released and is under active development.
+
+* Add type hints (:issue:`191`)
+
+* Add proper trove classifiers for PyPy support (:issue:`198`)
+
+* Scale back depending on ``ctypes`` for manylinux support detection (:issue:`171`)
+
+* Use ``sys.implementation.name`` where appropriate for ``packaging.tags`` (:issue:`193`)
+
+19.2 - 2019-09-18
+~~~~~~~~~~~~~~~~~
+
+* Remove dependency on ``attrs`` (:issue:`178`, :issue:`179`)
+
+* Use appropriate fallbacks for CPython ABI tag (:issue:`181`, :issue:`185`)
+
+* Add manylinux2014 support (:issue:`186`)
+
+* Improve ABI detection (:issue:`181`)
+
+* Properly handle debug wheels for Python 3.8 (:issue:`172`)
+
+* Improve detection of debug builds on Windows (:issue:`194`)
 
 19.1 - 2019-07-30
 ~~~~~~~~~~~~~~~~~